public_cloud/requirements-public.txt
langchain
openai
tiktoken
faiss-gpu
faiss-cpu
rouge-score
pandas
pyyaml
